+++ R/cumulativePercentage.R
@@ -1,12 +1,12 @@
-#' Plot the cumulative percentage tag allocation in sample
+#' Plot the cumulative percentage tag allocation
 #' 
 #' Plot the difference between the cumulative percentage tag allocation in 
 #' paired samples.
 #' 
 #' @param se An object of 
-#' \link[SummarizedExperiment]{RangedSummarizedExperiment}
-#' with assays of raw counts, ratios, background correct ratios,
-#' smoothed ratios and z-scores. It should be an element of output of 
+#' \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment}
+#' with assays of raw counts, transfomred ratios, background correct ratios,
+#' smoothed ratios and z-scores. It should be an element of the output of 
 #' \link{smoothRatiosByChromosome}.
 #' @param binWidth numeric(1) or integer(1). The width of each bin.
 #' @param backgroundCorrectionAssay character(1). Assays names 
@@ -31,9 +31,10 @@
 #'  PMCID: PMC3342857
 #' @examples 
 #' data(triplicates.counts)
+#' library(SummarizedExperiment)
 #' se <- triplicates.counts
 #' gps <- c("26", "28", "29")
-#' se <- log2se(se, 
+#' se <- log2se(se, transformation = "log2Ratio", 
 #'              nucleosomeCols = paste0("N", gps, ".bam"),
 #'              genomeCols = paste0("G", gps, ".bam"))
 #' se <- smoothRatiosByChromosome(se, chr="chr18")